完善商品添加功能, TODO 分布式事务

+++ product-common/src/main/java/io/github/thesixonenine/product/dto/Attr.java
@@ -16,7 +16,7 @@ public class Attr {
      * attrValue : 黑色
      */
 
-    private Integer attrId;
+    private Long attrId;
     private String attrName;
     private String attrValue;
 }
